refactor(tests): split test_enrichment_models.py into per-model modules
Break up the 883-line monolith into focused test files:
- conftest.py (shared org fixture)
- test_focus_area.py, test_skill.py, test_sdg.py
- test_operating_region.py, test_engagement_opportunity.py
- test_organization_enrichment.py, test_enum_schema_sync.py
- test_evaluation_enrichment.py
